LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/generated/svg-image.sub-expected.txt

Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY&body=%0A++++++%3Csvg+xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22+width%3D%22123%22+height%3D%22123%22%3E%0A++++++++%3Crect+fill%3D%22lime%22+width%3D%22123%22+height%3D%22123%22%2F%3E%0A++++++%3C%2Fsvg%3E%0A++++&mime=image%2Fsvg%2Bxml
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY&body=%0A++++++%3Csvg+xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22+width%3D%22123%22+height%3D%22123%22%3E%0A++++++++%3Crect+fill%3D%22lime%22+width%3D%22123%22+height%3D%22123%22%2F%3E%0A++++++%3C%2Fsvg%3E%0A++++&mime=image%2Fsvg%2Bxml
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY&body=%0A++++++%3Csvg+xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22+width%3D%22123%22+height%3D%22123%22%3E%0A++++++++%3Crect+fill%3D%22lime%22+width%3D%22123%22+height%3D%22123%22%2F%3E%0A++++++%3C%2Fsvg%3E%0A++++&mime=image%2Fsvg%2Bxml
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY&body=%0A++++++%3Csvg+xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22+width%3D%22123%22+height%3D%22123%22%3E%0A++++++++%3Crect+fill%3D%22lime%22+width%3D%22123%22+height%3D%22123%22%2F%3E%0A++++++%3C%2Fsvg%3E%0A++++&mime=image%2Fsvg%2Bxml

FAIL sec-fetch-site - Not sent to non-trustworthy same-origin destination no attributes assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-site - Not sent to non-trustworthy same-site destination no attributes promise_test: Unhandled rejection with value: object "[object Event]"
FAIL sec-fetch-site - Not sent to non-trustworthy cross-site destination no attributes assert_not_own_property: unexpected property "sec-fetch-site" is found on object
FAIL sec-fetch-mode - Not sent to non-trustworthy same-origin destination no attributes assert_not_own_property: unexpected property "sec-fetch-mode" is found on object
FAIL sec-fetch-mode - Not sent to non-trustworthy same-site destination no attributes promise_test: Unhandled rejection with value: object "[object Event]"
FAIL sec-fetch-mode - Not sent to non-trustworthy cross-site destination no attributes assert_not_own_property: unexpected property "sec-fetch-mode" is found on object
FAIL sec-fetch-dest - Not sent to non-trustworthy same-origin destination no attributes assert_not_own_property: unexpected property "sec-fetch-dest" is found on object
FAIL sec-fetch-dest - Not sent to non-trustworthy same-site destination no attributes promise_test: Unhandled rejection with value: object "[object Event]"
FAIL sec-fetch-dest - Not sent to non-trustworthy cross-site destination no attributes assert_not_own_property: unexpected property "sec-fetch-dest" is found on object
PASS sec-fetch-user - Not sent to non-trustworthy same-origin destination no attributes
FAIL sec-fetch-user - Not sent to non-trustworthy same-site destination no attributes promise_test: Unhandled rejection with value: object "[object Event]"
PASS sec-fetch-user - Not sent to non-trustworthy cross-site destination no attributes
FAIL sec-fetch-site - HTTPS downgrade (header not sent) no attributes assert_not_own_property: unexpected property "sec-fetch-site" is found on object
PASS sec-fetch-site - HTTPS upgrade no attributes
PASS sec-fetch-site - HTTPS downgrade-upgrade no attributes
